(#49) Refactor Analyser tests in to partial classes
